Only run piscina for large enough file sets
